refactor: rename description provider to context
Renamed `provider` to `context` in description-related components for consistency with established naming conventions. Updated imports, exports, file paths, and references to reflect this change across projects, workspaces, and tasks.
