+++ apps/nextjs/src/components/pages/my-tasks/components/tasks-header/sort-menu/sort-menu.tsx
@@ -1,6 +1,6 @@
 import { memo, useCallback, useMemo } from 'react';
 import { SortMenu as TasksHeaderSortMenu } from '@/components/features/tasks/tasks-header';
-import { useMyTasksContext } from '@/components/pages/my-tasks/providers/provider';
+import { useMyTasksContext } from '@/components/pages/my-tasks/providers/context';
 import {
   type TaskListSortStatusCodeValue,
   useMyTasksTaskListStatus,
